# apps/api/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/sessions/sign-out":{"post":{"operationId":"SessionsController_signOut","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignOutResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"security":[{"cwf_session":[]}],"summary":"Sign out of this device","tags":["Sessions"]}},"/api/v1/sign-ins":{"get":{"operationId":"AccountsController_pending","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PendingResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"security":[{"cwf_session":[]}],"summary":"Get pending sign-in information","tags":["Accounts"]}},"/api/v1/privacy-notice":{"get":{"operationId":"AccountsController_notice","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoticeResponseDto"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"security":[{"cwf_session":[]}],"summary":"Read the current privacy notice","tags":["Accounts"]}},"/api/v1/onboarding":{"post":{"operationId":"AccountsController_onboard","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"security":[{"cwf_session":[]}],"summary":"Complete onboarding and join","tags":["Accounts"]}},"/api/v1/sign-ins/cancel":{"post":{"operationId":"AccountsController_cancel","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"security":[{"cwf_session":[]}],"summary":"Cancel and remove pending sign-in","tags":["Accounts"]}},"/api/v1/members/me":{"get":{"operationId":"AccountsController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"security":[{"cwf_session":[]}],"summary":"Read your own member account","tags":["Accounts"]}},"/api/v1/auth/google/start":{"get":{"operationId":"GoogleController_start","parameters":[{"name":"returnTo","required":false,"in":"query","schema":{"example":"/circles/join/example","type":"string"}}],"responses":{"302":{"description":"Redirects to Google."}},"summary":"Start Google sign-in","tags":["Authentication"]}},"/api/v1/auth/google/callback":{"get":{"operationId":"GoogleController_callback","parameters":[],"responses":{"302":{"description":"Redirects to the saved path, onboarding, or Not open yet."},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"summary":"Complete Google sign-in","tags":["Authentication"]}},"/api/v1/health":{"get":{"description":"Returns the readiness of the API process.","operationId":"HealthController_getHealth","parameters":[],"responses":{"200":{"description":"The API process is ready.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponseDto"}}}},"429":{"description":"The API rate limit was exceeded.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}},"500":{"description":"The API could not complete the request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponseDto"}}}}},"summary":"Check API health","tags":["Health"]}}},"info":{"title":"CreditWithFriends API","description":"Versioned API for CreditWithFriends.","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"cwf_session":{"type":"apiKey","in":"cookie","name":"cwf_session"}},"schemas":{"SignOutDataDto":{"type":"object","properties":{"signedOut":{"type":"boolean","example":true}},"required":["signedOut"]},"SignOutResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/SignOutDataDto"},"error":{"type":"string","nullable":true,"enum":[null],"example":null}},"required":["success","data","error"]},"ApiFieldErrorDto":{"type":"object","properties":{"field":{"type":"string","example":"email","description":"Name of the invalid field."},"reason":{"type":"string","example":"invalid_email","description":"Stable validation reason."}},"required":["field","reason"]},"ApiErrorDetailsDto":{"type":"object","properties":{"fieldErrors":{"example":[{"field":"email","reason":"invalid_email"}],"type":"array","items":{"$ref":"#/components/schemas/ApiFieldErrorDto"}}}},"ApiErrorDto":{"type":"object","properties":{"errorId":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"code":{"type":"string","example":"NOT_FOUND"},"type":{"type":"string","example":"HttpError"},"message":{"type":"string","example":"The requested resource was not found."},"userMessage":{"type":"string","example":"We could not find what you requested."},"details":{"$ref":"#/components/schemas/ApiErrorDetailsDto"},"statusCode":{"type":"number","example":404},"correlationId":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"requestId":{"type":"string","nullable":true,"example":null},"environment":{"type":"string","example":"Local"},"timestampUtc":{"type":"string","example":"2026-09-27T12:00:00.000Z"}},"required":["errorId","code","type","message","userMessage","details","statusCode","correlationId","requestId","environment","timestampUtc"]},"ApiErrorResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":false},"data":{"type":"string","nullable":true,"enum":[null],"example":null},"error":{"$ref":"#/components/schemas/ApiErrorDto"}},"required":["success","data","error"]},"PendingDataDto":{"type":"object","properties":{"name":{"type":"string","example":"Arihant"},"email":{"type":"string","example":"owner@example.in"},"returnPath":{"type":"string","example":"/circles/join/example"}},"required":["name","email","returnPath"]},"PendingResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/PendingDataDto"},"error":{"type":"string","nullable":true,"enum":[null],"example":null}},"required":["success","data","error"]},"NoticeDataDto":{"type":"object","properties":{"version":{"type":"number","example":1},"text":{"type":"string","example":"Privacy notice text"}},"required":["version","text"]},"NoticeResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/NoticeDataDto"},"error":{"type":"string","nullable":true,"enum":[null],"example":null}},"required":["success","data","error"]},"OnboardingDto":{"type":"object","properties":{"displayName":{"type":"string","example":"Arihant"},"whatsappNumber":{"type":"string","example":"+919876543210"},"isAdultConfirmed":{"type":"boolean","example":true},"isConsentGiven":{"type":"boolean","example":true},"privacyNoticeVersion":{"type":"number","example":1}},"required":["displayName","whatsappNumber","isAdultConfirmed","isConsentGiven","privacyNoticeVersion"]},"MemberDataDto":{"type":"object","properties":{"id":{"type":"string","example":"01960463-1700-7000-8000-000000000001"},"displayName":{"type":"string","example":"Arihant"},"whatsappE164":{"type":"string","example":"+919876543210"},"googleEmail":{"type":"string","example":"owner@example.in"},"googleAccountId":{"type":"string","example":"112233445566778899"}},"required":["id","displayName","whatsappE164","googleEmail","googleAccountId"]},"MemberResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/MemberDataDto"},"error":{"type":"string","nullable":true,"enum":[null],"example":null}},"required":["success","data","error"]},"ActionDataDto":{"type":"object","properties":{"removed":{"type":"boolean","example":true}},"required":["removed"]},"ActionResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/ActionDataDto"},"error":{"type":"string","nullable":true,"enum":[null],"example":null}},"required":["success","data","error"]},"HealthDataDto":{"type":"object","properties":{"status":{"type":"string","example":"ok","description":"Indicates that the API process is ready."}},"required":["status"]},"HealthResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/HealthDataDto"},"error":{"type":"string","nullable":true,"enum":[null],"example":null}},"required":["success","data","error"]}}}}